refactor(mobile): split market assets hook
