+++ apps/ledger-live-mobile/src/mvvm/features/Market/screens/MarketScreen/useMarketAssets.ts
@@ -1,14 +1,5 @@
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-  type Dispatch,
-  type SetStateAction,
-} from "react";
+import { useCallback, useMemo } from "react";
 import { useMarketData } from "@ledgerhq/live-common/market/hooks/useMarketDataProvider";
-import type { MarketCurrencyData } from "@ledgerhq/live-common/market/utils/types";
-import type { Unit } from "@ledgerhq/types-cryptoassets";
 import { useLocale, useTranslation } from "~/context/Locale";
 import { useSelector } from "~/context/hooks";
 import { counterValueCurrencySelector } from "~/reducers/settings";
@@ -18,16 +9,27 @@ import type {
   MarketListSorting,
 } from "~/reducers/types";
 import type { MarketAssetDisplayData } from "LLM/components/AssetListItem";
-import { mapMarketCurrencyToDisplayData } from "../../utils/marketAssetDisplay";
 import {
   getMarketListDisplayRange,
   getMarketListOrder,
   getMarketListRequestRange,
 } from "./marketListFilters";
+import {
+  canReachEnd,
+  type EmptyState,
+  getEmptyState,
+  getMarketAssets,
+  getMarketDataForDisplay,
+  getMarketFilter,
+} from "./marketAssetsHelpers";
+import { useMarketAssetCategoryState } from "./useMarketAssetCategoryState";
+import {
+  getNextPaginationState,
+  type PaginationParams,
+  useMarketAssetPagination,
+} from "./useMarketAssetPagination";
 
 const PAGE_SIZE = 20;
-const STOCK_MARKET_FILTER = "stock";
-const EMPTY_MARKET_DATA: MarketCurrencyData[] = [];
 
 export type MarketAssetsParams = {
   search?: string;
@@ -36,28 +38,6 @@ export type MarketAssetsParams = {
   timeframe?: MarketListFilterTimeframe;
   starredMarketCoins?: string[];
 };
-
-type PaginationState = {
-  page: number;
-  search: string;
-  category: MarketListCategory;
-  favoriteIdsKey: string;
-  sorting: MarketListSorting;
-  timeframe: MarketListFilterTimeframe;
-};
-
-type PaginationParams = Omit<PaginationState, "page">;
-
-type CategoryState = {
-  isFavoritesCategory: boolean;
-  isStocksCategory: boolean;
-  sortedFavoriteIds: string[] | undefined;
-  favoriteIdsKey: string;
-  hasFavoriteIds: boolean;
-  shouldFetchAssets: boolean;
-};
-
-type EmptyState = "favorites" | "stocks" | undefined;
 
 export interface MarketAssetsResult {
   assets: MarketAssetDisplayData[];
@@ -87,7 +67,7 @@ export function useMarketAssets({
     favoriteIdsKey,
     hasFavoriteIds,
     shouldFetchAssets,
-  } = useMarketCategoryState({ category, normalizedSearch, starredMarketCoins });
+  } = useMarketAssetCategoryState({ category, normalizedSearch, starredMarketCoins });
   const paginationParams = useMemo<PaginationParams>(
     () => ({
       search: normalizedSearch,
@@ -169,194 +149,4 @@ export function useMarketAssets({
     emptyState: getEmptyState({ isFavoritesCategory, hasFavoriteIds, isStocksCategory }),
     onEndReached,
   };
-}
-
-function useMarketCategoryState({
-  category,
-  normalizedSearch,
-  starredMarketCoins,
-}: {
-  category: MarketListCategory;
-  normalizedSearch: string;
-  starredMarketCoins: string[];
-}): CategoryState {
-  const hasSearch = normalizedSearch.length > 0;
-  const isFavoritesCategory = !hasSearch && category === "starred";
-  const isStocksCategory = !hasSearch && category === "stocks";
-  const sortedFavoriteIds = useMemo(
-    () => getSortedFavoriteIds(isFavoritesCategory, starredMarketCoins),
-    [isFavoritesCategory, starredMarketCoins],
-  );
-  const favoriteIdsKey = sortedFavoriteIds?.join(",") ?? "";
-  const hasFavoriteIds = Boolean(sortedFavoriteIds?.length);
-  const shouldFetchAssets = !isFavoritesCategory || hasFavoriteIds;
-
-  return {
-    isFavoritesCategory,
-    isStocksCategory,
-    sortedFavoriteIds,
-    favoriteIdsKey,
-    hasFavoriteIds,
-    shouldFetchAssets,
-  };
-}
-
-function useMarketAssetPagination({
-  params,
-  shouldFetchAssets,
-}: {
-  params: PaginationParams;
-  shouldFetchAssets: boolean;
-}): {
-  page: number;
-  requestedPage: number;
-  setPagination: Dispatch<SetStateAction<PaginationState>>;
-} {
-  const [pagination, setPagination] = useState<PaginationState>(() =>
-    createPaginationState(params),
-  );
-  const isPaginationSynced = isSamePaginationTarget(pagination, params);
-  const page = isPaginationSynced ? pagination.page : 1;
-  const requestedPage = shouldFetchAssets ? page : 0;
-
-  useEffect(() => {
-    if (!isPaginationSynced) {
-      setPagination(createPaginationState(params));
-    }
-  }, [isPaginationSynced, params]);
-
-  return { page, requestedPage, setPagination };
-}
-
-function createPaginationState(params: PaginationParams): PaginationState {
-  return {
-    page: 1,
-    ...params,
-  };
-}
-
-function getNextPaginationState(
-  current: PaginationState,
-  params: PaginationParams,
-): PaginationState {
-  if (!isSamePaginationTarget(current, params)) {
-    return createPaginationState(params);
-  }
-
-  return { ...current, page: current.page + 1 };
-}
-
-function isSamePaginationTarget(state: PaginationState, params: PaginationParams): boolean {
-  return (
-    state.search === params.search &&
-    state.category === params.category &&
-    state.favoriteIdsKey === params.favoriteIdsKey &&
-    state.sorting === params.sorting &&
-    state.timeframe === params.timeframe
-  );
-}
-
-function getSortedFavoriteIds(
-  isFavoritesCategory: boolean,
-  starredMarketCoins: string[],
-): string[] | undefined {
-  if (!isFavoritesCategory) {
-    return undefined;
-  }
-
-  return [...starredMarketCoins].sort((firstId, secondId) => firstId.localeCompare(secondId));
-}
-
-function getMarketFilter(isStocksCategory: boolean): string | undefined {
-  return isStocksCategory ? STOCK_MARKET_FILTER : undefined;
-}
-
-function getMarketDataForDisplay(
-  data: MarketCurrencyData[],
-  shouldFetchAssets: boolean,
-): MarketCurrencyData[] {
-  return shouldFetchAssets ? data : EMPTY_MARKET_DATA;
-}
-
-function getMarketAssets({
-  marketData,
-  isStocksCategory,
-  counterCurrency,
-  counterValueUnit,
-  displayRange,
-  locale,
-  t,
-}: {
-  marketData: MarketCurrencyData[];
-  isStocksCategory: boolean;
-  counterCurrency: string;
-  counterValueUnit: Unit;
-  displayRange: Parameters<typeof mapMarketCurrencyToDisplayData>[1]["range"];
-  locale: string;
-  t: Parameters<typeof mapMarketCurrencyToDisplayData>[1]["t"];
-}): MarketAssetDisplayData[] {
-  const filteredMarketData = isStocksCategory
-    ? marketData.filter(isStockMarketCurrency)
-    : marketData;
-  const uniqueById = [...new Map(filteredMarketData.map(item => [item.id, item])).values()];
-
-  return uniqueById.map(item =>
-    mapMarketCurrencyToDisplayData(item, {
-      counterCurrency,
-      counterValueUnit,
-      range: displayRange,
-      locale,
-      t,
-    }),
-  );
-}
-
-function canReachEnd({
-  shouldFetchAssets,
-  canLoadMore,
-  loading,
-  isFetchingNextPage,
-  isError,
-}: {
-  shouldFetchAssets: boolean;
-  canLoadMore: boolean;
-  loading: boolean;
-  isFetchingNextPage: boolean;
-  isError: boolean;
-}): boolean {
-  return shouldFetchAssets && canLoadMore && !loading && !isFetchingNextPage && !isError;
-}
-
-function getEmptyState({
-  isFavoritesCategory,
-  hasFavoriteIds,
-  isStocksCategory,
-}: {
-  isFavoritesCategory: boolean;
-  hasFavoriteIds: boolean;
-  isStocksCategory: boolean;
-}): EmptyState {
-  if (isFavoritesCategory && !hasFavoriteIds) {
-    return "favorites";
-  }
-
-  if (isStocksCategory) {
-    return "stocks";
-  }
-
-  return undefined;
-}
-
-function isStockMarketCurrency(item: MarketCurrencyData): boolean {
-  const id = item.id.toLowerCase();
-  const name = item.name.toLowerCase();
-
-  return (
-    id.includes("xstock") ||
-    id.includes("tokenized-stock") ||
-    id.includes("prestocks") ||
-    name.includes("xstock") ||
-    name.includes("tokenized stock") ||
-    name.includes("prestocks")
-  );
 }
